refactor(L.7): migrate F0Curve/DSFile time types to TimePos and mHz
- F0Curve: timestep doubleTimePos(us), values vector<double>vector<int32_t>(mHz)
- DSFile: Phone/Note start/duration doubleTimePos, offset doubleTimePos
- DSFile I/O: load HzmHz via hzToMhzBatch, save mHzHz via mhzToHzBatch
- NoteCommands: NoteSnapshot::duration doubleTimePos
- PitchCommands: ModulationDriftCommand F0 vectors doubleint32_t
- PitchProcessor: use CurveTools::movingAverage/smoothstepCrossfade, adapt to mHz
- PianoRollRenderer: convert TimePossec at boundary, mHzMIDI for display
- PianoRollView/InputHandler: adapt to TimePos and int32_t F0 values
- PropertyPanel: adapt to TimePos/mHz types
- Tests: update TestF0Curve and TestPitchProcessor for new types

+++ src/apps/PitchLabeler/gui/PitchLabelerPage.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <dsfw/IPageActions.h>
+#include <dstools/TimePos.h>
 
 #include <QWidget>
 #include <QAction>
@@ -153,7 +154,7 @@ private:
 
     // State
     bool m_abComparisonActive = false;
-    std::vector<double> m_originalF0;
+    std::vector<int32_t> m_originalF0;
     QList<QShortcut *> m_windowShortcuts;
     QUndoStack *m_undoStack = nullptr;
     std::unique_ptr<QProxyStyle> m_jumpClickStyle;
